rename decisionSlice to treeSlice for clarity
