+++ src/components/Tree/Nodes/DefaultNode/DefaultNode.spec.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 import { cleanup, render, screen } from '@testing-library/react';
 import { DefaultNode } from 'components/Tree/Nodes/DefaultNode/DefaultNode';
 import { ReactFlowProvider } from 'reactflow';
-import { DecisionStatus } from 'store/DecisionSlice/decisionSlice';
+import { DecisionStatus } from 'store/TreeSlice/treeSlice';
 import { afterEach, describe, expect, test } from 'vitest';
 
 interface TestComponentProps {
