+++ src/store/TreeSlice/layout.spec.ts
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom';
 import { PositionUnawareDecisionTree } from 'store/DagNodeSlice/dagNodeSlice';
-import { layoutTree } from 'store/DecisionSlice/layout';
+import { layoutTree } from 'store/TreeSlice/layout';
 import { describe, expect, test } from 'vitest';
 
 describe('DAG layout', () => {
