+++ src/store/index.ts
@@ -1,11 +1,11 @@
 import { createDagEdgeSlice, DagEdgeSlice } from 'store/DagEdgeSlice/dagEdgeSlice';
 import { createDagNodeSlice, DagNodeSlice } from 'store/DagNodeSlice/dagNodeSlice';
-import { createDecisionSlice, DecisionSlice } from 'store/DecisionSlice/decisionSlice';
 import {
   createDecisionTreeStore,
   DecisionTreeStore,
 } from 'store/DecisionTreeStore/decisionTreeStore';
 import { createHelpSlice, HelpSlice } from 'store/HelpSlice/helpSlice';
+import { createTreeSlice, TreeSlice } from 'store/TreeSlice/treeSlice';
 import { create } from 'zustand';
 import { devtools } from 'zustand/middleware';
 
@@ -16,16 +16,16 @@ export type {
   ShowDagNodeOptions,
 } from 'store/DagNodeSlice/dagNodeSlice';
 
-export type { TreeDirection, DecisionTree } from 'store/DecisionSlice/decisionSlice';
+export type { TreeDirection, DecisionTree } from 'store/TreeSlice/treeSlice';
 
 const useTreeStore = create<
-  DagNodeSlice & DecisionSlice & DagEdgeSlice & DecisionTreeStore & HelpSlice
+  DagNodeSlice & TreeSlice & DagEdgeSlice & DecisionTreeStore & HelpSlice
 >()(
   devtools(
     (...args) => ({
       ...createDagNodeSlice(...args),
       ...createDagEdgeSlice(...args),
-      ...createDecisionSlice(...args),
+      ...createTreeSlice(...args),
       ...createDecisionTreeStore(...args),
       ...createHelpSlice(...args),
     }),
